faq: split answer and remove question tasks
